Refactor detailed list
